Rm mummichog
It's too delicate: the newest version is bugged and you have to use a downgraded version of networkx in order to not crash so I'd have to set up a requirements.txt and instructions to make the proper Python environment. It didn't return significant results anyway. If we want to mention it in the eventual paper, I can come back to clean it up and add it, but probably not a priority right now.